% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runs.tests.R
\name{ALRuns.plots}
\alias{ALRuns.plots}
\title{Function that takes the output from an assessment and produces plots of the runs test for the age and length composition.}
\usage{
ALRuns.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL,
  log.resid = TRUE
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{log.resid}{}
}
\description{
Function that takes the output from an assessment and produces plots of the runs test for the age and length composition.
}
