% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bound.vec.plots.R
\name{Bound.vec.plots}
\alias{Bound.vec.plots}
\title{Plots time and age vectors and their bounds for stock-assessment models}
\usage{
Bound.vec.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles 
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}
}
\value{
Graphics
}
\description{
Graphically displays estimated vectors and bound values for a stock-assessment model run.
}
\examples{
\donttest{
Bound.vec.plots(gag)
}

}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
