% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSR.time.plots.R
\name{BSR.time.plots}
\alias{BSR.time.plots}
\title{Biomass, spawner, and recruitmenet trajectories}
\usage{
BSR.time.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  start.drop = 0,
  graphics.type = NULL,
  use.color = TRUE,
  units.b = x$info$units.biomass,
  units.ssb = x$info$units.ssb,
  units.r = x$info$units.rec,
  legend.pos = "topright",
  from.zero = TRUE,
  BSR.references = list("Bmsy", "SSBmsy", "Rmsy")
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{start.drop}{Number of years at the start of the data to be omitted from
plots, as when a model includes an initialization period.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{units.b}{A text string (e.g. \code{"tons"} for labeling plots of stock biomass.}

\item{units.ssb}{A text string for labeling plots of spawning-stock biomass.}

\item{units.r}{a text string (e.g. \code{"million fish"}) for labeling plots of recruitment}

\item{legend.pos}{A text string compatible with the \code{legend} function of \code{R}.
Defines the position of the legend (ex. "bottomright", "bottom", etc.)}

\item{from.zero}{When \code{TRUE}, the Y-axis of each plot
(except recruitment deviations) starts at zero.}

\item{BSR.references}{A list of three character-string names to specify reference
points to appear on BSR plots, in order of biomass, spawning biomass, and recruits.
Input NULL for none, e.g., BSR.references=list(NULL,"SSBmsy", NULL).}
}
\value{
Graphics
}
\description{
The function \code{BSR.time.plots} generates time-series plots of stock biomass,
spawning stock (as biomass or egg production), and recruitment. Plots are made
on absolute scales and also scaled to reference points. Plots may include
reference lines at user-specified reference points.
}
\examples{
\donttest{
BSR.time.plots(gag)
}

}
\author{
M. Prager

E. Williams

K. Shertzer

R. Cheshire

K. Purcell
}
