% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CLD.total.plots.R
\name{CLD.total.plots}
\alias{CLD.total.plots}
\title{Time plots of model catch, landings, and discards by fishery}
\usage{
CLD.total.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, first.year = 0,
  units.CLD.n = x$info$units.numbers, units.CLD.w = x$info$units.biomass,
  plot.proportion = TRUE)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{first.year}{An integer (e.g. 1960) indicating the first year for plots.
The subsetting is done via row names of the matrices being plotted.}

\item{units.CLD.n}{A text string (e.g. \code{"million fish"}) used in labeling
Y-axis of plots of numbers caught, landed, or discarded.}

\item{units.CLD.w}{A text string (e.g. \code{"t"}) used in labeling the Y-axis
of plots of biomass caught, landed, or discarded.}

\item{plot.proportion}{When \code{TRUE}, additional plots are made, scaled as
proportions.}
}
\value{
None
}
\description{
The function \code{CLD.total.plots} generates barplots of estimated catch, landings,
and discards (in numbers and weight) over time. Each bar is subdivided by fishery.
}
\examples{
\donttest{
CLD.total.plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

