% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.plots.R
\name{Cohort.plots}
\alias{Cohort.plots}
\title{Plots of age-composition data observed and model predicted with color coding by cohort}
\usage{
Cohort.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}
}
\value{
Graphics
}
\description{
The function \code{Cohort.plots} generates bar plot of observed age composition data
with bars colored to track cohorts through time
and model predicted values indicated by black dots and connected lines.
Color coding tracks cohorts, accounting for skipped years;
colors may be recycled depending on the number of ages.
}
\examples{
\donttest{
Cohort.plots(gag)
}
}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
