% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comp.plots.archive.R
\name{Comp.plots.archive}
\alias{Comp.plots.archive}
\title{Plots of age- and length-composition fits}
\usage{
Comp.plots.archive(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, units = x$info$units.length,
  plot.options = FGGetOptions())
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{units}{a character string (e.g. \code{"cm"}) for labeling the X-axis of
length-composition plots.}
}
\value{
Graphics
}
\description{
The function \code{Comp.plots} generates bubble plots of residuals of age-
and length-composition fits for the entire time frame of the assessment.
Bubbles are scaled to the largest residual in each plot. So that the absolute
scale of deviations can be judged, a small inset plot displays the angular
deviation between observed and predicted values each year.
}
\examples{
\donttest{
Comp.plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

