% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comp.plots.color.cohort.R
\name{Comp.plots.color.cohort}
\alias{Comp.plots.color.cohort}
\title{Plots of age-composition data observed and model predicted with color coding by cohort}
\usage{
Comp.plots.color.cohort(x, DataName = deparse(substitute(x)),
  graphics.type = NULL)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}
}
\value{
Graphics
}
\description{
The function \code{Comp.plots.color.cohort} generates bar plot of observed age composition data
with bars colored to track cohorts through time
and model predicted values indicated by black dots and connected lines.
Skipped years are accounted for and color coding stays with cohort, then recycled based on number of ages.
}
\examples{
\donttest{
Comp.plots.color.cohort(gag)
}
}
\author{
Erik H. Williams
}

