% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Plots.r
\name{Data.Plots}
\alias{Data.Plots}
\title{Plot of available data from the output of an assessment
The function \code{Data.Plots} generates a timeseries plot of the different data sources. Optionally, the size of the points for the age, length, and weight composition can be proportion to the sample size of input data}
\usage{
Data.Plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL,
  use.color = TRUE,
  scaleSize = FALSE,
  cexval = 1,
  fill = TRUE,
  sampleName = "neff"
)
}
\arguments{
\item{x}{an R list with output from the assessment model.}

\item{DataName}{string used in plot titles. Defaults to argument \code{x}.
#' @param draft modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{scaleSize}{a boolean whether to scale size of point for age and length composition based on the sample size. Default is FALSE}

\item{fill}{a boolean whether to fill the point, this is useful for diagnosing how much overlap occurs in the points, Default if TRUE}

\item{sampleName}{a character string of the length and age composition name to search for in the tseries matrix. Do not include the '.'. Default is 'neff'}
}
\value{
Graphics
}
\description{
Plot of available data from the output of an assessment
The function \code{Data.Plots} generates a timeseries plot of the different data sources. Optionally, the size of the points for the age, length, and weight composition can be proportion to the sample size of input data
}
\examples{
\donttest{
Data.Plots(gag)
}
}
\author{
MT Vincent
}
