% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EqRec.plots.R
\name{EqRec.plots}
\alias{EqRec.plots}
\title{Equilibrium-recruitment plots}
\usage{
EqRec.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE,
  units.L = x$info$units.landings[1], units.SSB = x$info$units.ssb[1],
  units.B = x$info$units.biomass[1], units.D = x$info$units.discards[1],
  units.Y = x$info$units.yield[1], units.R = x$info$units.rec[1],
  user.Eq = NULL)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{units.L}{A character string (e.g. \code{"pounds"} for labeling the plot
of equilibrium landings.}

\item{units.SSB}{A character string for labeling the plot of equilibrium
spawning biomass.}

\item{units.B}{A character string used in labeling the plot of equilibrium biomass.}

\item{units.D}{A character string used in labeling the plot of equilibrium discards.}

\item{units.Y}{A character string used in labeling the plot of equilibrium yield.}

\item{units.R}{A character string used in labeling the plot of equilibrium recruitment.}

\item{units.Eq}{A list whose elements are names of additional columns of
\code{x$eq.series} to be plotted against \emph{F}.}
}
\value{
Graphics
}
\description{
The function \code{EqRec.plots} generates plots of quantities (calculated by the
user’s model) at equilibrium as a function of fishing mortality rate \emph{F}.
This function’s operation is almost identical to that of function \code{PerRec.plots}
. It exists to allow the analyst’s model to examine, in two potentially
different resolutions, quantities that vary with \emph{F}.
}
\examples{
\donttest{
BSR.time.plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

