% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.age.plot.R
\name{F.age.plot}
\alias{F.age.plot}
\title{The routine \code{F.age.plots} provides a time series plot of fishing mortality at age for all ages. Plots are of absolute estimated F at age.}
\usage{
F.age.plot(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL,
  use.color = TRUE,
  legend.pos = "topleft"
)
}
\arguments{
\item{x}{an R list with output from the assessment model}

\item{DataName}{string used in plot titles. Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report. When \code{FALSE} main titles are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved. When \code{NULL}, no plots are saved.}

\item{use.color}{binary to determine if use default R colors (FALSE) or create using rainbow and number of ages (TRUE). default=TRUE}

\item{legend.pos}{a text string compatible with the \code{legend} function of \code{R}. Defines the position of the legend (ex. "bottomright","bottom",etc.)}
}
\value{
Graphics
}
\description{
The routine \code{F.age.plots} provides a time series plot of fishing mortality at age for all ages. Plots are of absolute estimated F at age.
}
\examples{
\donttest{
F.age.plot(gag)
}
}
\author{
M Vincent
}
