% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/F.time.plots.R
\name{F.time.plots}
\alias{F.time.plots}
\title{Plots of fishing mortality rate over time}
\usage{
F.time.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  start.drop = 0, graphics.type = NULL, use.color = TRUE,
  legend.pos = "topleft", F.references = NULL,
  plot.options = FGGetOptions())
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{start.drop}{Number of years at the start of the data to be omitted from
plots, as when a model includes an initialization period.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{legend.pos}{A text string compatible with the \code{legend} function of \code{R}.
Defines the position of the legend (ex. "bottomright", "bottom", etc.)}

\item{F.references}{A list of character-string vectors to specify reference
points to appear on \emph{F} plots.}
}
\value{
Graphics
}
\description{
The routine \code{F.time.plots} provides time plots of fishing mortality rate
\emph{F} over time. Plots are of absolute estimated \emph{F} and \emph{F}
relative to various reference points. Through the argument \code{F.references},
arbitrary reference points can be included in plots, one by one or together.
}
\examples{
\donttest{
F.time.plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

