% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runs.tests.R
\name{FGRunsPlot}
\alias{FGRunsPlot}
\title{Simple plot function for runs test that plots the test for a single index or mean length/age time series. This function needs to be given the residuals precalculated for a single object with names Yr and residuals}
\usage{
FGRunsPlot(year, scaled.resids, lab.y, PlotTitle)
}
\arguments{
\item{year}{a vector with the year of the residual}

\item{scaled.resids}{a vector of the}

\item{lab.y}{a string of the y axis label}

\item{PlotTitle}{a string of the plot title}
}
\value{
Graphics
}
\description{
Simple plot function for runs test that plots the test for a single index or mean length/age time series. This function needs to be given the residuals precalculated for a single object with names Yr and residuals
}
\author{
M T Vincent (many ideas from Henning Winker and Laurence Kell)
}
