% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gag-data.R
\docType{data}
\name{gag}
\alias{gag}
\title{GAG data output from an assessment model}
\format{a list object}
\usage{
data(gag)
}
\description{
Data from a recent stock assessment of gag grouper.
Data have been read into an object named "gag" using the dget()
function on the R-formatted output of the assessment model.
}
\examples{
data(gag)
str(gag)
}
\references{
Prager et al. (2015) NOAA Technical Memorandum NMFS-SEFSC-XXX
(\href{http://spo.nmfs.noaa.gov/tm.htm}{NMFS SPO})
}
\keyword{dataset}

