% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.plots.R
\name{Growth.plots}
\alias{Growth.plots}
\title{Plots of fish size (and other quantities) at age}
\usage{
Growth.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL,
  use.color = TRUE,
  units.length = x$info$units.length,
  units.weight = x$info$units.weight,
  plot.all = FALSE
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles 
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}.}

\item{units.length}{Character string giving the units of measure associated
with the lengths at age found in \code{x$a.series$length}.}

\item{units.weight}{Character string giving the units of measure associated
with weights at age found in \code{x$a.series$weight}.}

\item{plot.all}{When \code{TRUE}, all columns in \code{x$a.series} are plotted
against \code{x$a.series$age}.}
}
\value{
Graphics
}
\description{
The function Growth.plots provides plots of length, weight, and other 
quantities at age. Length is also plotted with confidence intervals if the 
CV of length at age is found. There is support for models with more than one 
growth curve, such as models that describe growth as varying by sex.
}
\examples{
\donttest{
Growth.plots(gag)
}

}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
