% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Index.plots.archive.R
\name{Index.plots.archive}
\alias{Index.plots.archive}
\title{Plots of fits to abundance indices}
\usage{
Index.plots.archive(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, connect.obsd = FALSE,
  from.zero = TRUE, two.panel = TRUE, log.resid = TRUE,
  draw.barplot = TRUE, draw.pairs = TRUE, plot.options = FGGetOptions())
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{connect.obsd}{In plots of observed and predicted data, observed points
are joined by a line when \code{TRUE}.}

\item{from.zero}{When \code{TRUE}, the Y-axis of each plot
(except recruitment deviations) starts at zero.}

\item{two.panel}{When \code{TRUE}, the observed-predicted plot and residual
time plot are drawn (and saved) as two panels of a single figure.  When \code{FALSE},
that are drawn and saved independently.}

\item{log.resid}{When \code{TRUE} residuals are computed as \eqn{R = log(\frac{U}{\hat{U}})}.}

\item{draw.barplot}{When \code{TRUE}, a stacked barplot of standardized
residulas is drawn.}

\item{draw.pairs}{When \code{TRUE}, a scatterplot matrix of observed abundance
indices is drawn.}
}
\value{
Graphics
}
\description{
The routine \code{Index.plots} provides graphs for a visual evaluation of fits to
abundance indices. This includes graphs of observed and predicted indices
and graphs of residuals.
}
\examples{
\donttest{
Index.plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

