% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Index.plots.beta.r
\name{Index.plots.beta}
\alias{Index.plots.beta}
\title{Abundance Index plots from stock assessment models}
\usage{
Index.plots.beta(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, connect.obsd = FALSE,
  from.zero = TRUE, two.panel = TRUE, log.resid = FALSE)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{connect.obsd}{When \code{TRUE} a line connecting observed points is plotted}

\item{from.zero}{When \code{TRUE}, the Y-axis of each plot (except recruitment deviations) starts at zero.}

\item{two.panel}{when \code{TRUE}, the observed-predicted plot and residual time plot are drawn as two-panels.  When \code{FALSE} they are drawn independently.}

\item{log.resid}{When \code{TRUE} residuals are computed as \eqn{R = log(U \diagup\hat{U})}.}

\item{draw.barplot}{When \code{TRUE}, a stacked barplot of standardized residuals is drawn.}

\item{draw.pairs}{When \code{TRUE}, a scatterplot matrix of observed abundance indicies is drawn.}
}
\value{
Graphics
}
\description{
Plots observed and predicted and diagnostic plots for index development
}
\examples{
\donttest{
Index.plots.beta(gag)
}
}
\author{
E.H. Williams

M.H. Prager

A. Stevens
}

