% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landings.plots.R
\name{Landings.plots}
\alias{Landings.plots}
\title{Landings and discards trajectories}
\usage{
Landings.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, from.zero = TRUE,
  L.units = x$info$units.landings, D.units = x$info$units.discards,
  start.drop = 0, L.obs.pre = TRUE, D.obs.pre = TRUE)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}.}

\item{from.zero}{When \code{TRUE}, the Y-axis of each plot
(except recruitment deviations) starts at zero.}

\item{L.units}{Character vector containing units of measure associate with
each landings series found in \code{x$t.series}.}

\item{D.units}{Same as the preceding, but for discards.}

\item{L.obs.pre}{When \code{TRUE}, landings data are interpreted as pairs of
columns (observed and predicted). When \code{FALSE}, each column represents
observed landings.}

\item{D.obs.pre}{Same as the preceding, but for discards.}
}
\value{
Graphics
}
\description{
The function \code{Landings.plots} provides time trajectories of landings and
discards by fishery. Optional arguments specify whether data represent
observed and predicted landings or simply independent series of landings.
The same holds for discards.
}
\examples{
\donttest{
Landings.plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

