% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFZ.age.plots.R
\name{NFZ.age.plots}
\alias{NFZ.age.plots}
\title{At-age matrix plots}
\usage{
NFZ.age.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  start.drop = 0,
  graphics.type = NULL,
  use.color = TRUE,
  units.naa = x$info$units.naa,
  units.biomass = x$info$units.biomass,
  units.ssb = x$info$units.ssb,
  max.bub = 4,
  user.plots = NULL,
  plot.CLD = FALSE,
  drop.last.NB = FALSE
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{start.drop}{Number of years at the start of the data to be omitted from
plots, as when a model includes an initialization period.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{units.naa}{A text string (e.g. \code{"million fish"}) used in labeling
the plot of numbers at age.}

\item{units.biomass}{A text string (e.g. \code{"t"}) used in labeling the plot
of biomass at age.}

\item{max.bub}{A numerical scalar for the maximum bubble size (e.g. \code{"4.0"})
in bubble plots of numbers at age and biomass at age.}

\item{user.plots}{A vector of text strings with the names of additional matrix
components of \code{x} to be plotted.}

\item{plot.CLD}{When \code{TRUE}, each matrix in \code{x$CLD.est.mats} is plotted,
in addition to the plots described here.}

\item{drop.last.NB}{boolean for whether to drop the last year of the abundance and biomass at age, when \code{TRUE} doesn't plot the last year of abundance, default \code{TRUE}.}
}
\value{
Graphics
}
\description{
The function \code{NFZ.age.plots} generates barplots of estimated abundance and
mortality at age over time and bubble plots of number and biomass at age.  Bubble
areas are scaled to the largest value within each plot.
}
\examples{
\donttest{
NFZ.age.plots(gag)
}
}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
