% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parm.agevec.bounds.plots.r
\name{Parm.agevec.bounds.plots}
\alias{Parm.agevec.bounds.plots}
\title{Age vector parameter and bounds plots for stock-assessment models}
\usage{
Parm.agevec.bounds.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}
}
\value{
Graphics
}
\description{
Graphically displays age vector parameter and bound values for a stock-assessment model runs.
}
\examples{
\donttest{
Param.agevec.bounds.plots(gag)
}
}
\author{
Erik H. Williams

Kyle W. Shertzer
}

