% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Parm.bounds.plots.R
\name{Parm.bounds.plots}
\alias{Parm.bounds.plots}
\title{Parameter and bounds plots for stock-assessment models}
\usage{
Parm.bounds.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, bound.tol = 0.01)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{bound.tol}{the tolerance, as percentage of bound range, for flagging estimates close to bounds.}
}
\value{
Graphics
}
\description{
Graphically displays parameter and bound values for a stock-assessment model runs.
}
\examples{
\donttest{
Param.bounds.plots(gag)
}
}
\author{
Erik H. Williams

KM Purcell
}

