% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parm.plots.R
\name{Parm.plots}
\alias{Parm.plots}
\title{Graphically displays parameter and bound values for a stock-assessment model runs.}
\usage{
Parm.plots(
  x,
  DataName = deparse(substitute(x)),
  graphics.type = NULL,
  bound.tol = 0.01
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{bound.tol}{the tolerance, as percentage of bound range, for flagging estimates close to bounds.}
}
\value{
Graphics
}
\description{
Graphically displays parameter and bound values for a stock-assessment model runs.
}
\examples{
\donttest{
Parm.plots(gag)
}

}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
