% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerRec.plots.R
\name{PerRec.plots}
\alias{PerRec.plots}
\title{Per-recruit plots}
\usage{
PerRec.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, units.ypr = x$info$units.ypr,
  user.PR = NULL, legend.pos = "topright", F.references = NULL)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{units.ypr}{A character string (e.g. \code{"pounds"}) used in labeling
the plot of yield per recruit.}

\item{user.PR}{A list whose elements are names of additional columns of
\code{x$pr.series} to be plotted against \emph{F}.}

\item{legend.pos}{A text string compatible with the \code{legend} function of \code{R}.
Defines the position of the legend (ex. "bottomright", "bottom", etc.)}

\item{F.references}{A list of character-string vectors to specify reference
points to appear on \emph{F} plots.}
}
\value{
Graphics
}
\description{
The function \code{PerRec.plots} generates plots of quantities
on a per-recruit basis as a function of fishing mortality
rate \emph{F}.
}
\examples{
\donttest{
PerRec.plots(gag)
}
}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}

