% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase.plots.R
\name{Phase.plots}
\alias{Phase.plots}
\title{Plots of fishing and spawning biomass status}
\usage{
Phase.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL,
  use.color = TRUE,
  start.drop = 0,
  end.drop = 0,
  year.pos = 1,
  from.zero = TRUE,
  legend.pos = NULL,
  Xaxis.F = TRUE,
  F.series = "F.full",
  B.series = "SSB",
  F.B.references = list("Fmsy", "msst")
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}.}

\item{start.drop}{Number of years at the start of the data to be omitted from
plots, as when a model includes an initialization period.}

\item{end.drop}{Number of years at the end of data to be omitted from
plots, as when a model includes a projection period.}

\item{year.pos}{An integer (= 1, 2, 3, or 4) defining the position of text relative to points.
The text indicates first and last years in the time series. A value of year.pos=0
turns off the text feature.}

\item{from.zero}{When \code{TRUE}, the Y-axis of each plot
starts at zero.}

\item{legend.pos}{A text string compatible with the \code{legend} function of \code{R}.
Defines the position of the legend (ex. "bottomright", "bottom", etc.). The default
NULL places text on the plot itself.}

\item{Xaxis.F}{When \code{TRUE}, F is plotted as the X axis, otherwise, B.}

\item{F.series}{Character string specifying the name of the F metric to be plotted}

\item{B.series}{Character string specifying the name of the B (or SSB) metric to be plotted}

\item{F.B.references}{a list of two character-sting names, the first specifying the
fishing reference point and the second the biomass reference point. Insert
a value of NULL if reference points are not desired for plotting.}
}
\value{
Graphics
}
\description{
The function Phase.plots provides plots of fishing status versus spawning biomass status.
}
\examples{
\donttest{
Phase.plots(gag)
}
}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
