% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runs.tests.R
\name{runs.sig3}
\alias{runs.sig3}
\title{Function to do runs.test and calculate 3 sigma limits
runs test is conducted with library(randtests)}
\usage{
runs.sig3(x, type = "resid", mixing = "positive.correlated")
}
\arguments{
\item{x}{residuals from CPUE fits}

\item{type}{only c("resid","observations")}

\item{mixing}{c("less","greater","two.sided"). Default less is checking for postive autocorrelation only}
}
\value{
runs test p value and 3 x sigma limits
}
\description{
Function to do runs.test and calculate 3 sigma limits
runs test is conducted with library(randtests)
}
\author{
Henning Winker (JRC-EC) and Laurence Kell (Sea++) Edited by Matt Vincent (NOAA SEFSC)
}
