% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Selectivity.plots.R
\name{Selectivity.plots}
\alias{Selectivity.plots}
\title{Estimated selectivity curves}
\usage{
Selectivity.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, plot.points = TRUE,
  units.length = x$info$units.length, units.age = x$info$units.age)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{plot.points}{A logical values; if \code{TRUE}, points will be plotted on curves.}

\item{units.length}{A text string (e.g. \code{"mm"}) used in labeling plots
of selectivity at length.}

\item{units.age}{A text string (e.g. \code{"years"}) used in labeling plots
of selectivity at age.}
}
\value{
Graphics
}
\description{
The function \code{Selectivity.plots} generates plots of selectivity curves
at age and length.
}
\examples{
\donttest{
Selectivity.plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

