% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comp.plots.R
\name{Comp.plots}
\alias{Comp.plots}
\title{Plots of age- and length-composition fits}
\usage{
Comp.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, units = x$info$units.length,
  p.resid = FALSE, corr = TRUE, c.min = 0.25, max.bub = 8)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{units}{a character string (e.g. \code{"cm"}) for labeling the X-axis of
length-composition plots.}

\item{p.resid}{when FALSE, scaled to area of residuals, otherwise scaled to area of the
Pearson residuals (multinomial).}

\item{corr}{when \code{FALSE}, angular deviation is displayed in the inset plot; otherwise,
correlations.}

\item{c.min}{lower bound on the y-axis range of the inset plot, applies only when corr is \code{TRUE}}

\item{max.bub}{cex value for maximum bubble size, default is 8.0.}
}
\value{
Graphics
}
\description{
The function \code{Comp.plots} generates bubble plots of residuals of age-
and length-composition fits for the entire time frame of the assessment.
Bubble areas are scaled to the largest residual within each plot. Optionally,
a small inset plot displays the correlation or angular
deviation between observed and predicted values each year. In addition,
\code{Comp.plots} generates plots of predicted and observed mean compositions
pooled across years.
}
\examples{
\donttest{
Comp.plots(gag)
}
}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}

