% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comp.yearly.plots.R
\name{Comp.yearly.plots}
\alias{Comp.yearly.plots}
\title{Plots of age- and length-composition fits by year}
\usage{
Comp.yearly.plots(x, DataName = deparse(substitute(x)), draft = TRUE,
  graphics.type = NULL, use.color = TRUE, units = x$info$units.length,
  print.angle = !compact, print.n = !compact, print.neff = !compact,
  plot.neff = TRUE, compute.neff = FALSE, print.year = compact,
  compact = FALSE, uniform = TRUE, connect.obsd = FALSE,
  plot.options = FGGetOptions())
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{units}{a character string (e.g. \code{"cm"}) for labeling the X-axis of
length-composition plots.}

\item{print.angle}{When \code{TRUE}, angular deviation between predicted and
observed composition is printed in the plot.}

\item{print.n}{When \code{TRUE}, \emph{N} is printed in each plot.}

\item{print.neff}{When \code{TRUE}, \emph{N_eff} is printed in each plot.}

\item{plot.neff}{When \code{TRUE}, a timeplot of \emph{N_eff} is made for each data
series}

\item{compute.neff}{When \code{TRUE}, \strong{FishGraph} computes \emph{N_eff}
internally, rather than using values stored by the user.  The computations are
based on the sample sizes and observed and estimated proportions.}

\item{print.year}{When \code{TRUE}, the year is printed in each plot.}

\item{compact}{When \code{TRUE}, plots are arranged in 7x3 matrix on each page.}

\item{uniform}{When \code{TRUE}, all years of the same data series are scaled
the same.}
}
\value{
Graphics
}
\description{
The routine \code{Comp.yearly.plots} generates plots of age- and length-composition
fits by year and data series. Optionally, the sample size N, effective sample
size Neff, angular deviation can be printed on the plot surface. Plots of the
ratio Neff/N over time may be made. Plots are available in two formats: one
plot per page, or a compact format with 21 plots per page.
}
\examples{
\donttest{
Comp_yearly_plots(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

