% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdat2ascii.R
\name{rdat2ascii}
\alias{rdat2ascii}
\title{Rdat conversion program}
\usage{
rdat2ascii(x, file = paste(deparse(substitute(x)), ".txt", sep = ""))
}
\arguments{
\item{x}{an R list with output from the assessment models.}
}
\value{
ASCII file
}
\description{
The function \code{rdat2ascii} converts an \code{.Rdat} object into an ASCII
file for alternative uses.
}
\examples{
\donttest{
rdat2ascii(gag)
}
}
\author{
M. Prager

Erik H. Williams

Andi Stephens

Kyle W. Shertzer
}

