% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivity.plots.R
\name{Selectivity.plots}
\alias{Selectivity.plots}
\title{Estimated selectivity curves}
\usage{
Selectivity.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL,
  use.color = TRUE,
  plot.points = TRUE,
  units.length = x$info$units.length,
  units.age = x$info$units.age,
  compact = FALSE,
  legend.pos = "right"
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles 
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{plot.points}{A logical values; if \code{TRUE}, points will be plotted on curves.}

\item{units.length}{A text string (e.g. \code{"mm"}) used in labeling plots
of selectivity at length.}

\item{units.age}{A text string (e.g. \code{"years"}) used in labeling plots 
of selectivity at age.}

\item{compact}{When \code{TRUE}, time-varying selectivity curves are placed on a single plot per fishery.}

\item{legend.pos}{A text string compatible with the \code{legend} function of \code{R}.
Defines the position of the legend (ex. "bottomright", "bottom", etc.). Applies only if 
\code{compact = TRUE}.}
}
\value{
Graphics
}
\description{
The function \code{Selectivity.plots} generates plots of selectivity curves 
at age and length.
}
\examples{
\donttest{
Selectivity.plots(gag)
}

}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
