% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockRec.plots.R
\name{StockRec.plots}
\alias{StockRec.plots}
\title{Stock-recruitment plots}
\usage{
StockRec.plots(
  x,
  DataName = deparse(substitute(x)),
  draft = TRUE,
  graphics.type = NULL,
  use.color = TRUE,
  start.drop = 0,
  units.ssb = x$info$units.ssb,
  units.rec = x$info$units.rec,
  rec.model = x$info$rec.model,
  draw.model = TRUE,
  draw.lowess = FALSE,
  draw.time = TRUE,
  year.pos = 1
)
}
\arguments{
\item{x}{an R list with output from the assessment models.}

\item{DataName}{string used in plot titles.  Defaults to argument \code{x}.}

\item{draft}{modifies plots for use in a report.  When \code{FALSE} main titles 
are omitted.}

\item{graphics.type}{a vector of graphics file types to which graphics are saved.
When \code{NULL}, no plots are saved.}

\item{use.color}{plots are made in grayscale when \code{FALSE}}

\item{start.drop}{Number of years at the start of the data to be omitted from
plots, as when a model includes an initialization period.}

\item{units.ssb}{A text string (e.g. \code{"tons"} or \code{"10^9 eggs"}) used
in labeling plots of spawning stock.}

\item{units.rec}{A text string (e.g. \code{"million fish"}) for labeling plots 
of recruitment}

\item{rec.model}{Specifies the type of recruitment function to draw.  Valid 
values are: \code{"BH", "BH-steep", "Ricker", "Ricker-steep", "Mean"}.}

\item{draw.model}{If \code{TRUE}, a function curve is drawn on plots of stock
and recruitment.}

\item{draw.lowess}{If \code{TRUE}, a lowess smooth is drawn on plots of stock
and recruitment.}

\item{draw.time}{If \code{TRUE}, stock-recruitment points are connected to indicate
progression of time.}

\item{year.pos}{An integer (= 1, 2, 3, or 4) defining the position of text relative to points. 
The text indicates first and last years in the time series, and applies only if draw.time=TRUE. A
value of year.pos=0 turns off the text feature.}
}
\value{
Graphics
}
\description{
The function \code{StockRec.plots} generates plots of stock vs. recruitment in 
linear and logarithmic scales.
}
\examples{
\donttest{
StockRec.plots(gag)
}

}
\author{
M Prager

E Williams

K Shertzer

R Cheshire

K Purcell
}
